divide tests of task1 functions into different files
